Agressive simplification for cs contexts
Modification of the methods Context.{left/right}Simplify to eliminate priority level that cannot be broken using the simple shallow priority semantics.
However, the original shallow-priority semantics is too strong, and we use here a "direct priority conflicts" detection instead of the "shallow priority conflicts".
This is done inside the method Production.shallowConflicts: if the priority policy is set to DEEP, the method automatically call Production.directConflicts instead.