Refactor CharacterClass: use ICharacterClass where possible
The main rationale behind this refactoring is that the class
`CharacterClassSymbol` (a subclass of `Symbol`) was being "abused".
It was used for instance in the lookahead restriction of symbols
and reduce actions and in `State.doReduces`/`State.addReduceAction`.
Those use sites do not need to know anything related to a grammar
symbol, but really only need to know which characters are in the CC.
Therefore, it is better to directly use the `ICharacterClass` interface
at these places. In order to accomplish this, the following changes
have been made:

- `union`, `intersection`, and `difference` used to be called on the
  `CharacterClassFactory`, and `isEmpty` could only be called on
  `CharacterClassSymbol`. Now, they can be called directly on
  `ICharacterClass`.
- `toAterm` used to be a method in the `CharacterClassSymbol`.
  It is now also a method in `ICharacterClass`, where a more specific
  implementation for the "single" and "range" classes (resolving the
  `FIXME` comment).
  - This adds a dependency from `tableinterfaces` to
    `org.metaborg.terms`.

Note that the `CharacterClass` symbol class has been renamed to
`CharacterClassSymbol`, to reduce the confusion between the grammar
symbol and the `ICharacterClass` data structure.

As a final remark: my autoformatter has sometimes reformatted spacing
and import groups, even in classes where I only changed one line.
Unfortunately, this does generate some noise in the diff.