Always, automatically, generated namespaced version of the grammar
Before this commit you would need to call a menu item manually to generate a namespaced grammar file. This was done because the generated file would be put into the syntax directory. Those were put there because the files needed to have the SDF3 on-save handler called on them as well.
With these changes, that menu is called on-save. The generated file will end up in src-gen instead of syntax. The on-save handler generates all other files for the namespaced grammar as it generates the namespaced grammar.